requirements.txt
streamlit
openai
PyMuPDF
Pillow
python-dotenv
crewai[tools]